+++ src/main/java/logisticspipes/proxy/computers/objects/CCItemIdentifierBuilder.java
@@ -19,7 +19,7 @@ public class CCItemIdentifierBuilder implements ILPCCTypeHolder {
     private int itemData = 0;
 
     @CCCommand(description = "Set the itemID for this ItemIdentifierBuilder")
-    public void setItemID(Double id) {
+    public void setItemID(Long id) {
         itemID = id.intValue();
         itemIDName = null;
     }
@@ -39,7 +39,7 @@ public class CCItemIdentifierBuilder implements ILPCCTypeHolder {
     }
 
     @CCCommand(description = "Set the item data/damage for this ItemIdentifierBuilder")
-    public void setItemData(Double data) {
+    public void setItemData(Long data) {
         itemData = data.intValue();
     }
