+++ src/main/java/logisticspipes/proxy/computers/objects/CCSinkResponder.java
@@ -57,8 +57,8 @@ public class CCSinkResponder implements ILPCCTypeHolder {
 
     @CCCommand(
             description = "Sends the response to the CC QuickSort module to accept the sink for the givven amount with the givven priority")
-    public void acceptSink(Double amount, Double priority) {
-        canSink = ((Double) (amount > 0 ? amount : 0D)).intValue();
+    public void acceptSink(Long amount, Long priority) {
+        canSink = ((Long) (amount > 0 ? amount : 0L)).intValue();
         this.priority = priority.intValue();
         done = true;
     }
